corgs -> cogs
